+++ tests/data/test_data.py
@@ -148,7 +148,7 @@ class TestData:
     @property
     def visualize_2_pse(self):
         return f"{self.test_data_repo}/analysis/1SUO.xtal.{self.visualize_2_design_case}.pze"
-    
+
     visualize_3_profile_type: str = ""
     visualize_3_score_reversed: bool = False
     visualize_3_design_case: str = "csv_vis"
@@ -157,7 +157,18 @@ class TestData:
     @property
     def visualize_3_pse(self):
         return f"{self.test_data_repo}/analysis/1SUO.xtal.{self.visualize_3_design_case}.pze"
-    
+
+    # visualize_4_profile_type: str = ""
+    # visualize_4_score_reversed: bool = False
+    visualize_4_design_case: str = "csv_vis_grouped"
+    visualize_4_group_label: str = "group"
+    visualize_4_mutant_label: str = "mutant"
+    visualize_4_score_label: str = "normalized"
+
+    # visualize_4_use_global_score: bool = True
+    @property
+    def visualize_4_pse(self):
+        return f"{self.test_data_repo}/analysis/1SUO.xtal.{self.visualize_4_design_case}.pze"
 
     @property
     def entropy_best_hits(self):
@@ -275,6 +286,7 @@ class KeyData:
     minimum_mutant_file: str = '../tests/data/mutagenese/evaluate_pssm_ent_surf.mannual.mut.txt'
     ddg_file: str = '../tests/data/pytia_ddg/1SUO_pred_mask.csv'
     visualize_csv: str = '../tests/data/csv/experimetal_data_mock.csv'
+    visualize_csv_grouped: str = '../tests/data/csv/experimetal_data_mock_label.csv'
     evaluate_pse_path: str = None  # type: ignore
     gremlin_pkl_fp_homomer: str = None  # type: ignore