+++ CHANGELOG.md
@@ -18,6 +18,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Utils: 
+  - `require_not_none`:
+    - `fallback_setup`:  Function to call if the attribute is None
+
+### Changed
+
+- Tab `Visualize`: mutant loading from CSV files with WT labels
+- UI: `lineEdit_group_name` -> `comboBox_group_name` to support csv col name
+- `MutantVisualizer`: allow `run_mutagenesis_tasks` to handel multi-branch mutant tree
+- Test: 
+  - `TestWorker`: allow `check_existed_mutant_tree` to return existing mutant tree if the tree check passed
+  - tests on visualizing csv with group label
+  
+### Fixed
+
+- Tests: 
+  - Teardown: `SingletonAbstract` subclass resets
+    - to avoid instance reused across multiple tests
+
+### Removed
+
 ## [1.7.12] - 2025-01-02
 
 ### Added